rename Move to IMove, add drawableObject method and BorderBoard extends Brique

+++ IMove.java
@@ -3,6 +3,7 @@ public interface IMove{
 	public boolean collision(BoardOject object);
 
   public void setMove(double xMove, double yMove);
-	public int getXMove();
-	public int getYMove();
+	
+	public double getXMove();
+	public double getYMove();
 }
